requirements.txt
# Core OAuth Automation Dependencies
selenium>=4.15.0
requests>=2.31.0
webdriver-manager>=4.0.0

# Web Framework for API/Service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Database and Storage
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Background Tasks and Scheduling
celery>=5.3.0
redis>=5.0.0
apscheduler>=3.10.0

# Logging and Monitoring
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# HTTP and API Tools
httpx>=0.25.0
aiofiles>=23.2.0

# Date and Time Handling
python-dateutil>=2.8.0
pytz>=2023.3

# Security and Authentication
cryptography>=41.0.0
jwt>=1.3.1
passlib[bcrypt]>=1.7.4

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Cloud Deployment
gunicorn>=21.2.0
whitenoise>=6.6.0

# System and Process Management
psutil>=5.9.0

# Optional: Enhanced error handling
tenacity>=8.2.0 